Refactored float32 functions to asm/f32
