+++ asm/f32/axpy.go
@@ -4,29 +4,29 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package asm
+package f32
 
-func SaxpyUnitary(alpha float32, x, y []float32) {
+func AxpyUnitary(alpha float32, x, y []float32) {
 	for i, v := range x {
 		y[i] += alpha * v
 	}
 }
 
-func SaxpyUnitaryTo(dst []float32, alpha float32, x, y []float32) {
+func AxpyUnitaryTo(dst []float32, alpha float32, x, y []float32) {
 	for i, v := range x {
 		dst[i] = alpha*v + y[i]
 	}
 }
 
-func SaxpyInc(alpha float32, x, y []float32, n, incX, incY, ix, iy uintptr) {
+/*func AxpyInc(alpha float32, x, y []float32, n, incX, incY, ix, iy uintptr) {
 	for i := 0; i < int(n); i++ {
 		y[iy] += alpha * x[ix]
 		ix += incX
 		iy += incY
 	}
-}
+}*/
 
-func SaxpyIncTo(dst []float32, incDst, idst uintptr, alpha float32, x, y []float32, n, incX, incY, ix, iy uintptr) {
+func AxpyIncTo(dst []float32, incDst, idst uintptr, alpha float32, x, y []float32, n, incX, incY, ix, iy uintptr) {
 	for i := 0; i < int(n); i++ {
 		dst[idst] = alpha*x[ix] + y[iy]
 		ix += incX
