+++ asm/f32/dot.go
@@ -4,18 +4,18 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package asm
+package f32
 
-func DsdotUnitary(x, y []float32) (sum float64) {
+func DotUnitary(x, y []float32) (sum float32) {
 	for i, v := range x {
-		sum += float64(y[i]) * float64(v)
+		sum += y[i] * v
 	}
 	return
 }
 
-func DsdotInc(x, y []float32, n, incX, incY, ix, iy uintptr) (sum float64) {
+func DotInc(x, y []float32, n, incX, incY, ix, iy uintptr) (sum float32) {
 	for i := 0; i < int(n); i++ {
-		sum += float64(y[iy]) * float64(x[ix])
+		sum += y[iy] * x[ix]
 		ix += incX
 		iy += incY
 	}
