+++ asm/f32/scal.go
@@ -4,22 +4,22 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package asm
+package f32
 
-func SscalUnitary(alpha float32, x []float32) {
+func ScalUnitary(alpha float32, x []float32) {
 	for i := range x {
 		x[i] *= alpha
 	}
 }
 
-func SscalUnitaryTo(dst []float32, alpha float32, x []float32) {
+func ScalUnitaryTo(dst []float32, alpha float32, x []float32) {
 	for i, v := range x {
 		dst[i] = alpha * v
 	}
 }
 
 // incX must be positive.
-func SscalInc(alpha float32, x []float32, n, incX uintptr) {
+func ScalInc(alpha float32, x []float32, n, incX uintptr) {
 	var ix uintptr
 	for i := 0; i < int(n); i++ {
 		x[ix] *= alpha
@@ -28,7 +28,7 @@ func SscalInc(alpha float32, x []float32, n, incX uintptr) {
 }
 
 // incDst and incX must be positive.
-func SscalIncTo(dst []float32, incDst uintptr, alpha float32, x []float32, n, incX uintptr) {
+func ScalIncTo(dst []float32, incDst uintptr, alpha float32, x []float32, n, incX uintptr) {
 	var idst, ix uintptr
 	for i := 0; i < int(n); i++ {
 		dst[idst] = alpha * x[ix]
