+++ asm/f32/single_precision
